% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickVP.R
\name{quickVP}
\alias{quickVP}
\title{Automatically batch generate confirmations of participation}
\usage{
quickVP(dat, path, study_name, study_mail, study_adress, vp_hours)
}
\arguments{
\item{dat}{Dataframe, with rows doc_id, student_id, mail, date (see details)}

\item{path}{Character, path to the folder that will store resulting PDFs}

\item{study_name}{Character, name or title of the study}

\item{study_mail}{Character, mail adress of the investigator}

\item{study_adress}{Character vector, up to 10 lines of contact information
of the investigator}

\item{vp_hours}{Numeric, vp-hours awarded to the participant}
}
\description{
Automatically batch generate confirmations of participation
}
\details{
doc_id is a numeric id that is unique to the confirmation. The id is a
light protection against forgery.

student_id is numeric id unique to the student. This could, for instance, be
a matriculation number.

mail is a character that contains a mail adress of the participant.

date is a character that gives the data of participation.

You may set values of these columns to NA. Then, the respective section of
the confirmation will be skipped.
}
